INFUND-9494 update new note controller

+++ ifs-web-service/ifs-project-setup-mgt-service/src/main/java/org/innovateuk/ifs/project/notes/controller/FinanceChecksNotesAddNoteController.java
@@ -2,6 +2,7 @@ package org.innovateuk.ifs.project.notes.controller;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import org.innovateuk.ifs.application.service.OrganisationService;
+import org.innovateuk.ifs.commons.error.exception.ForbiddenActionException;
 import org.innovateuk.ifs.commons.rest.ValidationMessages;
 import org.innovateuk.ifs.commons.service.ServiceResult;
 import org.innovateuk.ifs.controller.ValidationHandler;
@@ -13,7 +14,6 @@ import org.innovateuk.ifs.project.financecheck.FinanceCheckService;
 import org.innovateuk.ifs.project.notes.form.FinanceChecksNotesAddNoteForm;
 import org.innovateuk.ifs.project.notes.form.FinanceChecksNotesFormConstraints;
 import org.innovateuk.ifs.project.notes.viewmodel.FinanceChecksNotesAddNoteViewModel;
-import org.innovateuk.ifs.project.queries.viewmodel.FinanceChecksQueriesAddQueryViewModel;
 import org.innovateuk.ifs.project.resource.ProjectResource;
 import org.innovateuk.ifs.user.resource.OrganisationResource;
 import org.innovateuk.ifs.user.resource.UserResource;
@@ -56,6 +56,7 @@ public class FinanceChecksNotesAddNoteController {
     private static final String FORM_COOKIE = "finance_checks_notes_new_note_form";
     private static final String FORM_ATTR = "form";
     private static final String NEW_NOTE_VIEW = "project/financecheck/new-note";
+    private static final String ORIGIN_GET_COOKIE = "finance_checks_notes_new_note_origin";
 
     @Autowired
     private OrganisationService organisationService;
@@ -81,10 +82,15 @@ public class FinanceChecksNotesAddNoteController {
                               HttpServletRequest request,
                               HttpServletResponse response) {
 
-        List<Long> attachments = loadAttachmentsFromCookie(request, projectId, organisationId);
-        model.addAttribute("model", populateNoteViewModel(projectId, organisationId, attachments));
-        model.addAttribute(FORM_ATTR, loadForm(request, projectId, organisationId).orElse(new FinanceChecksNotesAddNoteForm()));
-        return NEW_NOTE_VIEW;
+        if (projectService.getPartnerOrganisationsForProject(projectId).stream().filter(o -> o.getId() == organisationId).count() > 0) {
+            saveOriginCookie(response, projectId, organisationId, loggedInUser.getId());
+            List<Long> attachments = loadAttachmentsFromCookie(request, projectId, organisationId);
+            model.addAttribute("model", populateNoteViewModel(projectId, organisationId, attachments));
+            model.addAttribute(FORM_ATTR, loadForm(request, projectId, organisationId).orElse(new FinanceChecksNotesAddNoteForm()));
+            return NEW_NOTE_VIEW;
+        } else {
+            throw new ForbiddenActionException();
+        }
     }
 
     @PreAuthorize("hasPermission(#projectId, 'ACCESS_FINANCE_CHECKS_NOTES_SECTION')")
@@ -98,42 +104,46 @@ public class FinanceChecksNotesAddNoteController {
                            @ModelAttribute("loggedInUser") UserResource loggedInUser,
                            HttpServletRequest request,
                            HttpServletResponse response) {
-        Supplier<String> failureView = () -> {
-            List<Long> attachments = loadAttachmentsFromCookie(request, projectId, organisationId);
-            FinanceChecksNotesAddNoteViewModel viewModel = populateNoteViewModel(projectId, organisationId, attachments);
-            model.addAttribute("model", viewModel);
-            model.addAttribute(FORM_ATTR, form);
-            return NEW_NOTE_VIEW;
-        };
+        if (postParametersMatchOrigin(request, projectId, organisationId, loggedInUser.getId())) {
+            Supplier<String> failureView = () -> {
+                List<Long> attachments = loadAttachmentsFromCookie(request, projectId, organisationId);
+                FinanceChecksNotesAddNoteViewModel viewModel = populateNoteViewModel(projectId, organisationId, attachments);
+                model.addAttribute("model", viewModel);
+                model.addAttribute(FORM_ATTR, form);
+                return NEW_NOTE_VIEW;
+            };
 
-        return validationHandler.failNowOrSucceedWith(failureView, () -> {
+            return validationHandler.failNowOrSucceedWith(failureView, () -> {
 
-            ValidationMessages validationMessages = new ValidationMessages(bindingResult);
+                ValidationMessages validationMessages = new ValidationMessages(bindingResult);
 
-            ProjectFinanceResource projectFinance = projectFinanceService.getProjectFinance(projectId, organisationId);
+                ProjectFinanceResource projectFinance = projectFinanceService.getProjectFinance(projectId, organisationId);
 
-            List<AttachmentResource> attachmentResources = new ArrayList<>();
-            List<Long> attachments = loadAttachmentsFromCookie(request, projectId, organisationId);
-            attachments.forEach(attachment -> {
-                ServiceResult<AttachmentResource> fileEntry = financeCheckService.getAttachment(attachment);
-                if (fileEntry.isSuccess()) {
-                    attachmentResources.add(fileEntry.getSuccessObject());
-                }
+                List<AttachmentResource> attachmentResources = new ArrayList<>();
+                List<Long> attachments = loadAttachmentsFromCookie(request, projectId, organisationId);
+                attachments.forEach(attachment -> {
+                    ServiceResult<AttachmentResource> fileEntry = financeCheckService.getAttachment(attachment);
+                    if (fileEntry.isSuccess()) {
+                        attachmentResources.add(fileEntry.getSuccessObject());
+                    }
+                });
+
+                PostResource post = new PostResource(null, loggedInUser, form.getNote(), attachmentResources, ZonedDateTime.now());
+
+                List<PostResource> posts = new ArrayList<>();
+                posts.add(post);
+                NoteResource note = new NoteResource(null, projectFinance.getId(), posts, form.getNoteTitle(), ZonedDateTime.now());
+                ServiceResult<Long> result = financeCheckService.saveNote(note);
+                validationHandler.addAnyErrors(result);
+                return validationHandler.addAnyErrors(validationMessages, fieldErrorsToFieldErrors(), asGlobalErrors()).
+                        failNowOrSucceedWith(failureView, () -> {
+                            deleteCookies(response, projectId, organisationId);
+                            return redirectToNotePage(projectId, organisationId);
+                        });
             });
-
-            PostResource post = new PostResource(null, loggedInUser, form.getNote(), attachmentResources, ZonedDateTime.now());
-
-            List<PostResource> posts = new ArrayList<>();
-            posts.add(post);
-            NoteResource note = new NoteResource(null, projectFinance.getId(), posts, form.getNoteTitle(), ZonedDateTime.now());
-            ServiceResult<Long> result = financeCheckService.saveNote(note);
-            validationHandler.addAnyErrors(result);
-            return validationHandler.addAnyErrors(validationMessages, fieldErrorsToFieldErrors(), asGlobalErrors()).
-                    failNowOrSucceedWith(failureView, () -> {
-                        deleteCookies(response, projectId, organisationId);
-                        return redirectToNotePage(projectId, organisationId);
-                    });
-        });
+        } else {
+            throw new ForbiddenActionException();
+        }
     }
 
     @PreAuthorize("hasPermission(#projectId, 'ACCESS_FINANCE_CHECKS_NOTES_SECTION')")
@@ -147,27 +157,32 @@ public class FinanceChecksNotesAddNoteController {
                                         @ModelAttribute("loggedInUser") UserResource loggedInUser,
                                         HttpServletRequest request,
                                         HttpServletResponse response) {
-        List<Long> attachments = loadAttachmentsFromCookie(request, projectId, organisationId);
-        Supplier<String> onSuccess = () -> redirectTo(rootView(projectId, organisationId));
-        Supplier<String> onError = () -> {
-            model.addAttribute("model", populateNoteViewModel(projectId, organisationId, attachments));
-            model.addAttribute("form", form);
-            return NEW_NOTE_VIEW;
-        };
-        
-        return validationHandler.performActionOrBindErrorsToField("attachment", onError, onSuccess, () -> {
-            MultipartFile file = form.getAttachment();
 
-            ServiceResult<AttachmentResource> result = financeCheckService.uploadFile(projectId, file.getContentType(), file.getSize(), file.getOriginalFilename(), getMultipartFileBytes(file));
-            result.ifSuccessful(uploadedAttachment -> {
-                attachments.add(uploadedAttachment.id);
-                saveAttachmentsToCookie(response, attachments, projectId, organisationId);
-                saveFormToCookie(response, projectId, organisationId, form);
+        if (postParametersMatchOrigin(request, projectId, organisationId, loggedInUser.getId())) {
+            List<Long> attachments = loadAttachmentsFromCookie(request, projectId, organisationId);
+            Supplier<String> onSuccess = () -> redirectTo(rootView(projectId, organisationId));
+            Supplier<String> onError = () -> {
+                model.addAttribute("model", populateNoteViewModel(projectId, organisationId, attachments));
+                model.addAttribute("form", form);
+                return NEW_NOTE_VIEW;
+            };
+
+            return validationHandler.performActionOrBindErrorsToField("attachment", onError, onSuccess, () -> {
+                MultipartFile file = form.getAttachment();
+
+                ServiceResult<AttachmentResource> result = financeCheckService.uploadFile(projectId, file.getContentType(), file.getSize(), file.getOriginalFilename(), getMultipartFileBytes(file));
+                result.ifSuccessful(uploadedAttachment -> {
+                    attachments.add(uploadedAttachment.id);
+                    saveAttachmentsToCookie(response, attachments, projectId, organisationId);
+                    saveFormToCookie(response, projectId, organisationId, form);
+                });
+
+                model.addAttribute("model", populateNoteViewModel(projectId, organisationId, attachments));
+                return result;
             });
-
-            model.addAttribute("model", populateNoteViewModel(projectId, organisationId, attachments));
-            return result;
-        });
+        } else {
+            throw new ForbiddenActionException();
+        }
     }
 
     @PreAuthorize("hasPermission(#projectId, 'ACCESS_FINANCE_CHECKS_NOTES_SECTION')")
@@ -179,22 +194,25 @@ public class FinanceChecksNotesAddNoteController {
                                                          @PathVariable Long attachmentId,
                                                          @ModelAttribute("loggedInUser") UserResource loggedInUser,
                                                          HttpServletRequest request) {
-        List<Long> attachments = loadAttachmentsFromCookie(request, projectId, organisationId);
-        Optional<ByteArrayResource> content = Optional.empty();
-        Optional<FileEntryResource> fileDetails = Optional.empty();
+        if (projectService.getPartnerOrganisationsForProject(projectId).stream().filter(o -> o.getId() == organisationId).count() > 0) {
+            List<Long> attachments = loadAttachmentsFromCookie(request, projectId, organisationId);
+            Optional<ByteArrayResource> content = Optional.empty();
+            Optional<FileEntryResource> fileDetails = Optional.empty();
 
-        if (attachments.contains(attachmentId)) {
-            ServiceResult<Optional<ByteArrayResource>> fileContent = financeCheckService.downloadFile(attachmentId);
-            if (fileContent.isSuccess()) {
-                content = fileContent.getSuccessObject();
+            if (attachments.contains(attachmentId)) {
+                ServiceResult<Optional<ByteArrayResource>> fileContent = financeCheckService.downloadFile(attachmentId);
+                if (fileContent.isSuccess()) {
+                    content = fileContent.getSuccessObject();
+                }
+                ServiceResult<FileEntryResource> fileInfo = financeCheckService.getAttachmentInfo(attachmentId);
+                if (fileInfo.isSuccess()) {
+                    fileDetails = Optional.of(fileInfo.getSuccessObject());
+                }
             }
-            ServiceResult<FileEntryResource> fileInfo = financeCheckService.getAttachmentInfo(attachmentId);
-            if (fileInfo.isSuccess()) {
-                fileDetails = Optional.of(fileInfo.getSuccessObject());
-            }
-
+            return returnFileIfFoundOrThrowNotFoundException(content, fileDetails);
+        } else {
+            throw new ForbiddenActionException();
         }
-        return returnFileIfFoundOrThrowNotFoundException(content, fileDetails);
     }
 
     @PreAuthorize("hasPermission(#projectId, 'ACCESS_FINANCE_CHECKS_NOTES_SECTION')")
@@ -209,15 +227,19 @@ public class FinanceChecksNotesAddNoteController {
                                    HttpServletRequest request,
                                    HttpServletResponse response,
                                    Model model) {
-        List<Long> attachments = loadAttachmentsFromCookie(request, projectId, organisationId);
-        if (attachments.contains(attachmentId)) {
-            financeCheckService.deleteFile(attachmentId)
-                    .andOnSuccess(() -> attachments.remove(attachments.indexOf(attachmentId)));
-        }
-        saveAttachmentsToCookie(response, attachments, projectId, organisationId);
-        saveFormToCookie(response, projectId, organisationId, form);
+        if (postParametersMatchOrigin(request, projectId, organisationId, loggedInUser.getId())) {
+            List<Long> attachments = loadAttachmentsFromCookie(request, projectId, organisationId);
+            if (attachments.contains(attachmentId)) {
+                financeCheckService.deleteFile(attachmentId)
+                        .andOnSuccess(() -> attachments.remove(attachments.indexOf(attachmentId)));
+            }
+            saveAttachmentsToCookie(response, attachments, projectId, organisationId);
+            saveFormToCookie(response, projectId, organisationId, form);
 
-        return redirectTo(rootView(projectId, organisationId));
+            return redirectTo(rootView(projectId, organisationId));
+        } else {
+            throw new ForbiddenActionException();
+        }
     }
 
     @PreAuthorize("hasPermission(#projectId, 'ACCESS_FINANCE_CHECKS_NOTES_SECTION')")
@@ -227,10 +249,14 @@ public class FinanceChecksNotesAddNoteController {
                                 @ModelAttribute("loggedInUser") UserResource loggedInUser,
                                 HttpServletRequest request,
                                 HttpServletResponse response) {
-        List<Long> attachments = loadAttachmentsFromCookie(request, projectId, organisationId);
-        attachments.forEach(financeCheckService::deleteFile);
-        deleteCookies(response, projectId, organisationId);
-        return redirectToNotePage(projectId, organisationId);
+        if (projectService.getPartnerOrganisationsForProject(projectId).stream().filter(o -> o.getId() == organisationId).count() > 0) {
+            List<Long> attachments = loadAttachmentsFromCookie(request, projectId, organisationId);
+            attachments.forEach(financeCheckService::deleteFile);
+            deleteCookies(response, projectId, organisationId);
+            return redirectToNotePage(projectId, organisationId);
+        } else {
+            throw new ForbiddenActionException();
+        }
     }
 
     private FinanceChecksNotesAddNoteViewModel populateNoteViewModel(Long projectId, Long organisationId, List<Long> attachmentFileIds) {
@@ -287,6 +313,7 @@ public class FinanceChecksNotesAddNoteController {
     private void deleteCookies(HttpServletResponse response, Long projectId, Long organisationId) {
         cookieUtil.removeCookie(response, getCookieName(projectId, organisationId));
         cookieUtil.removeCookie(response, getFormCookieName(projectId, organisationId));
+        cookieUtil.removeCookie(response, ORIGIN_GET_COOKIE);
     }
 
     private String getFormCookieName(Long projectId, Long organisationId) {
@@ -309,5 +336,16 @@ public class FinanceChecksNotesAddNoteController {
 
     private void saveFormToCookie(HttpServletResponse response, Long projectId, Long organisationId, FinanceChecksNotesAddNoteForm form) {
         cookieUtil.saveToCookie(response, getFormCookieName(projectId, organisationId), JsonUtil.getSerializedObject(form));
+    }
+
+    private void saveOriginCookie(HttpServletResponse response, Long projectId, Long organisationId, Long userId) {
+        String jsonState = JsonUtil.getSerializedObject(Arrays.asList(projectId, organisationId, userId));
+        cookieUtil.saveToCookie(response, ORIGIN_GET_COOKIE, jsonState);
+    }
+
+    private boolean postParametersMatchOrigin(HttpServletRequest request, Long projectId, Long organisationId, Long userId){
+        List<Long> getParams = cookieUtil.getCookieAsList(request, ORIGIN_GET_COOKIE, new TypeReference<List<Long>>() {
+        });
+        return getParams.size() == 3 && getParams.get(0) == projectId && getParams.get(1) == organisationId && getParams.get(2) == userId;
     }
 }
